moved meeting-create component to calendar module. Moved related models from shared to app models folder
