+++ eta-company-calendar-frontend/src/app/calendar/calendar.module.ts
@@ -5,10 +5,15 @@ import { CalendarRoutingModule } from './calendar-routing.module';
 import { CalendarComponent } from './components/calendar.component';
 
 import { FullCalendarModule } from '@fullcalendar/angular'; // for FullCalendar!
+import { MeetingCreateComponent } from './modals/meeting-create.component';
 
 @NgModule({
   declarations: [
     CalendarComponent,
+    MeetingCreateComponent,
+  ],
+  entryComponents: [
+    MeetingCreateComponent,
   ],
   imports: [
     SharedModule,
