+++ eta-company-calendar-frontend/src/app/models/location.model.ts
