+++ eta-company-calendar-frontend/src/app/models/meeting-detail.model.ts
